feat: enhance save_scraper_data_pdf function with improved error handling and metadata storage

+++ src/apps/shared/utils/functions.py
@@ -314,34 +314,47 @@ def save_scraper_data_pdf(all_scraper, url, sobrenombre, collection, fs):
             file.write(all_scraper)
 
         with open(file_path, "rb") as file_data:
-            object_id = fs.put(file_data, filename=os.path.basename(file_path))
+            file_content = file_data.read()  
+        object_id = fs.put(file_content, 
+                           filename=os.path.basename(file_path), 
+                           metadata={
+                               "url": url,
+                               "sobrenombre": sobrenombre,
+                               "fecha_scraper": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
+                               "content": all_scraper  
+                           })
 
-            data = {
+        logger.info(f"Archivo subido a MongoDB con ObjectId: {object_id}")
+
+        if not object_id:
+            raise Exception("Error al guardar el archivo en GridFS, ObjectID no generado.")
+        data = {
                 "Objeto": object_id,
                 "Tipo": "Documento",
                 "Url": url,
-                "content": all_scraper,
                 "Fecha_scraper": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
                 "Etiquetas": ["planta", "plaga"],
             }
+            
+        collection.insert_one(data)
+        logger.info(f"Datos guardados en MongoDB para la URL: {url}")
 
-            collection.insert_one(data)
-            logger.info(f"Datos guardados en MongoDB para la URL: {url}")
+        delete_old_documents(url, collection, fs)
 
-            delete_old_documents(url, collection, fs)
-
-            response_data = {
+        response_data = {
                 "Tipo": "Documento",
                 "Url": url,
                 "Fecha_scraper": data["Fecha_scraper"],
                 "Etiquetas": data["Etiquetas"],
                 "Mensaje": "Los datos han sido scrapeados correctamente.",
             }
+        logger.info(f"DEBUG - Tipo de respuesta de save_scraper_data_pdf: {type(response_data)}")
+
 
         return response_data
     except Exception as e:
         logger.error(f"Error al guardar datos del scraper: {str(e)}")
-        raise
+        return {"error": f"Error al guardar datos del scraper: {str(e)}"}
 
 def process_scraper_data(all_scraper, url, sobrenombre, collection, fs):
     logger = get_logger("PROCESANDO DATOS DE ALL SCRAPER")
